+++ Controller/QueueController.php
@@ -2,6 +2,9 @@
 
 namespace Dtc\QueueBundle\Controller;
 
+use Dtc\QueueBundle\Doctrine\DoctrineJobManager;
+use Dtc\QueueBundle\Exception\UnsupportedException;
+use Dtc\QueueBundle\Model\BaseJob;
 use Dtc\QueueBundle\Model\Worker;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Method;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
@@ -36,6 +39,8 @@ class QueueController extends Controller
      * List jobs in system by default.
      *
      * @Route("/jobs_all", name="dtc_queue_jobs_all")
+     *
+     * @throws UnsupportedException
      */
     public function jobsAllAction()
     {
@@ -53,31 +58,106 @@ class QueueController extends Controller
     /**
      * @Route("/archive", name="dtc_queue_archive")
      * @Method({"POST"})
+     *
+     * @throws UnsupportedException
      */
     public function archiveAction(Request $request)
     {
+        return $this->streamResults($request, 'archiveAllJobs');
+    }
+
+    /**
+     * @Route("/reset-stalled", name="dtc_queue_reset_stalled")
+     *
+     * @param Request $request
+     *
+     * @return StreamedResponse
+     *
+     * @throws UnsupportedException
+     */
+    public function resetStalledAction(Request $request)
+    {
+        return $this->streamResults($request, 'resetStalledJobs');
+    }
+
+    /**
+     * @Route("/prune-stalled", name="dtc_queue_prune_stalled")
+     *
+     * @param Request $request
+     *
+     * @return StreamedResponse
+     *
+     * @throws UnsupportedException
+     */
+    public function pruneStalledAction(Request $request)
+    {
+        return $this->streamResults($request, 'pruneStalledJobs');
+    }
+
+    /**
+     * @param Request $request
+     * @param $functionName
+     *
+     * @return StreamedResponse
+     *
+     * @throws UnsupportedException
+     */
+    protected function streamResults(Request $request, $functionName)
+    {
+        $jobManager = $this->get('dtc_queue.manager.job');
+        if (!$jobManager instanceof DoctrineJobManager) {
+            throw new UnsupportedException('$jobManager must be instance of '.DoctrineJobManager::class);
+        }
+
+        $streamingResponse = new StreamedResponse($this->getStreamFunction($request, $functionName));
+        $streamingResponse->headers->set('Content-Type', 'application/x-ndjson');
+        $streamingResponse->headers->set('X-Accel-Buffering', 'no');
+
+        return $streamingResponse;
+    }
+
+    /**
+     * @param Request $request
+     * @param string  $functionName
+     *
+     * @return \Closure
+     */
+    protected function getStreamFunction(Request $request, $functionName)
+    {
+        $jobManager = $this->get('dtc_queue.manager.job');
         $workerName = $request->get('workerName');
         $methodName = $request->get('method');
+        $total = null;
+        $callback = function ($count, $totalCount) use (&$total) {
+            if (null !== $totalCount && null === $total) {
+                $total = $totalCount;
+                echo json_encode(['total' => $total]);
+                echo "\n";
+                flush();
 
-        $jobManager = $this->get('dtc_queue.manager.job');
-        $callback = function ($count) {
+                return;
+            }
             echo json_encode(['count' => $count]);
             echo "\n";
             flush();
         };
 
-        $streamingResponse = new StreamedResponse(function () use ($jobManager, $callback, $workerName, $methodName) {
-            $total = $jobManager->countLiveJobs($workerName, $methodName);
-            echo json_encode(['total' => $total]);
-            echo "\n";
-            flush();
-            if ($total > 0) {
-                $jobManager->archiveAllJobs($workerName, $methodName, $callback);
+        return function () use ($jobManager, $callback, $workerName, $methodName, $functionName, &$total) {
+            switch ($functionName) {
+                case 'archiveAllJobs':
+                    $total = $jobManager->countLiveJobs($workerName, $methodName);
+                    echo json_encode(['total' => $total]);
+                    echo "\n";
+                    flush();
+                    if ($total > 0) {
+                        $jobManager->archiveAllJobs($workerName, $methodName, $callback);
+                    }
+                    break;
+                default:
+                    $jobManager->$functionName($workerName, $methodName, $callback);
+                    break;
             }
-        });
-        $streamingResponse->headers->set('Content-Type', 'application/x-ndjson');
-
-        return $streamingResponse;
+        };
     }
 
     /**
@@ -85,6 +165,8 @@ class QueueController extends Controller
      *
      * @Template("@DtcQueue/Queue/jobs.html.twig")
      * @Route("/jobs", name="dtc_queue_jobs")
+     *
+     * @throws UnsupportedException
      */
     public function jobsAction()
     {
@@ -98,6 +180,7 @@ class QueueController extends Controller
         $this->addCssJs($params);
 
         $params['worker_methods'] = $this->get('dtc_queue.manager.job')->getWorkersAndMethods();
+        $params['prompt_message'] = 'This will archive all non-running jobs';
 
         return $params;
     }
@@ -118,6 +201,9 @@ class QueueController extends Controller
         $renderer->bind($gridSource);
         $params = $renderer->getParams();
         $this->addCssJs($params);
+
+        $params['worker_methods'] = $this->get('dtc_queue.manager.job')->getWorkersAndMethods(BaseJob::STATUS_RUNNING);
+        $params['prompt_message'] = 'This will prune all stalled jobs';
 
         return $params;
     }